fix trailing white space at end of first line of encoding causing decoder to hang, get rid of many files for encoder with makefile
