--- Makefile
@@ -1,2 +1,0 @@
-encode: encode.c tests.c
-	gcc -o encode encode.c tests.c
