+++ test/test_encoding_map.sh
@@ -1,8 +1,5 @@
 #!/usr/bin/env bash
-cd ..
-make
-cd test
-cp ../encode a.out
+gcc ../encode.c
 
 # Ensure that test_encoding_map() is uncommented in encode.c and
 # no text is printed to stdout
