--- tests.h
@@ -1,4 +1,0 @@
-void test_char_count(int char_count[]);
-void test_encoding_map(char *encoding[]);
-void print_expected_huffman_code(char *infile, char *encoding[]);
-void test_huffman_code(char *encoding[]);
